Merge the pbf-write branch that makes the output of splitter to be in the pbf format.
Use --output=xml to use the old xml format (and be aware that this is
version 0.5 of the osm xml format which is not widely accepted by
other tools).

Use --output=pbf to force pbf output (although that is the default
anyway).

Thanks to Francisco Moraes for the work on this feature.



git-svn-id: https://svn.mkgmap.org.uk/mkgmap/splitter@178 838fabe5-b1a8-4b81-9896-7a84386980e9