if keep-complete is active, make sure that small administrative boundaries are kept complete.
This option is used for relations with tags type=boundary & boundary=administrative & admin_level=*
where admin_level is equal to or higher than a given threshold.
The threshold value can be changed with the new option --wanted-admin-level, the default is 5.
Example usage: --keep-complete --wanted-admin-level=6
will make sure that all administrative boundaries with admin_level 
between 6 and 11 (including) are kept complete.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/splitter@423 838fabe5-b1a8-4b81-9896-7a84386980e9