improve calculation of split with option --num-tiles
Sometimes splitter did not find a split with exactly the wanted number of tiles. Hope this fixes it for reasonably small values in the option. If the wanted number is very high this might take long, if it is too high splitter will print a message and end with rc != 0.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/splitter@619 838fabe5-b1a8-4b81-9896-7a84386980e9